MOD: `ConvenioRepository` refatorado para alterar o tipo de ID de `Long` para `Integer`, garantindo consistência com a definição da entidade

+++ src/main/java/com/tlk/br/api/repositories/ConvenioRepository.java
@@ -1,10 +1,9 @@
 package com.tlk.br.api.repositories;
 
+import com.tlk.br.api.domain.entities.Convenio;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.tlk.br.api.domain.entities.Convenio;
 @Repository
-public interface ConvenioRepository extends JpaRepository<Convenio, Long> {
-
+public interface ConvenioRepository extends JpaRepository<Convenio, Integer> {
 }
